Fix constraints
