+++ GoInvest/Modules/Sources/QuoteDetail/QuoteDetailView.swift
@@ -1,0 +1,163 @@
+import UIKit
+
+class QuoteDetailView: UIView {
+    private let graphView: UIView = {
+        var view = UIView()
+        view.backgroundColor = .green
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+
+    private let buttonView: UIView = {
+        var view = UIView()
+        view.backgroundColor = .blue
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+
+    private let lastDateLabel = UILabel()
+    private let lastDate = UILabel()
+
+    private let closePriceLabel = UILabel()
+    private let closePriceAmountLabel = UILabel()
+
+    private let openPriceLabel = UILabel()
+    private let openPriceAmountLabel = UILabel()
+
+    private let averagePriceLabel = UILabel()
+    private let averagePriceAmountLabel = UILabel()
+
+    private let mainStackView = UIStackView()
+    private let labelStackView = UIStackView()
+    private let dateStackView = UIStackView()
+    private let openPriceStackView = UIStackView()
+    private let closePriceStackView = UIStackView()
+    private let averagePriceStackView = UIStackView()
+
+    private let addToPortfolioButton: UIButton = {
+        var button = UIButton()
+        button.backgroundColor = .black
+        button.translatesAutoresizingMaskIntoConstraints = false
+        button.layer.cornerRadius = 10
+        button.setTitle("Add to Portfolio", for: .normal)
+        button.setTitleColor(.white, for: .normal)
+        button.titleLabel?.font = .systemFont(ofSize: 19, weight: .semibold)
+        button.titleLabel?.adjustsFontForContentSizeCategory = true
+        return button
+    }()
+
+    init() {
+        super.init(frame: .zero)
+        self.translatesAutoresizingMaskIntoConstraints = false
+        setupUI()
+        setupLayout()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private func setupUI() {
+        applyStyleForLabel(for: lastDateLabel, text: "Date")
+        applyStyleForLabel(for: closePriceLabel, text: "Close price")
+        applyStyleForLabel(for: openPriceLabel, text: "Open price")
+        applyStyleForLabel(for: averagePriceLabel, text: "Average price")
+        applyStyleForAmountLabel(for: lastDate, text: "10.10.1010")
+        applyStyleForAmountLabel(for: closePriceAmountLabel, text: "1000 $")
+        applyStyleForAmountLabel(for: openPriceAmountLabel, text: "1000 $")
+        applyStyleForAmountLabel(for: averagePriceAmountLabel, text: "1000 $")
+    }
+
+    private func setupLayout() {
+        arrangeStackView(
+            for: dateStackView,
+            subviews: [lastDateLabel, lastDate]
+        )
+        lastDateLabel.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+        lastDate.setContentHuggingPriority(.defaultLow, for: .horizontal)
+        arrangeStackView(
+            for: closePriceStackView,
+            subviews: [closePriceLabel, closePriceAmountLabel]
+        )
+        closePriceLabel.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+        closePriceAmountLabel.setContentHuggingPriority(.defaultLow, for: .horizontal)
+        arrangeStackView(
+            for: openPriceStackView,
+            subviews: [openPriceLabel, openPriceAmountLabel]
+        )
+        openPriceLabel.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+        openPriceAmountLabel.setContentHuggingPriority(.defaultLow, for: .horizontal)
+        arrangeStackView(
+            for: averagePriceStackView,
+            subviews: [averagePriceLabel, averagePriceAmountLabel]
+        )
+        averagePriceLabel.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+        averagePriceAmountLabel.setContentHuggingPriority(.defaultLow, for: .horizontal)
+        arrangeStackView(
+            for: labelStackView,
+            subviews: [dateStackView,
+                       closePriceStackView,
+                       openPriceStackView,
+                       averagePriceStackView
+                       ],
+            spacing: 10.0,
+            axis: .vertical
+        )
+        arrangeStackView(
+            for: mainStackView,
+            subviews: [graphView,
+                       buttonView,
+                       labelStackView,
+                       addToPortfolioButton
+                       ],
+            spacing: 20,
+            axis: .vertical
+        )
+        self.addSubview(mainStackView)
+        NSLayoutConstraint.activate([
+            graphView.heightAnchor.constraint(equalToConstant: 300),
+            buttonView.heightAnchor.constraint(equalToConstant: 70),
+            addToPortfolioButton.heightAnchor.constraint(equalToConstant: 55),
+            mainStackView.topAnchor.constraint(equalTo: self.topAnchor),
+            mainStackView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+            mainStackView.leadingAnchor.constraint(equalTo: self.leadingAnchor)
+        ])
+    }
+}
+
+// MARK: - Apply style to UI Elements
+private extension QuoteDetailView {
+    func applyStyleForLabel(
+        for label: UILabel,
+        text: String) {
+            label.text = text
+            label.font = UIFont.systemFont(ofSize: 17, weight: .light)
+        }
+
+    func applyStyleForAmountLabel(
+        for label: UILabel,
+        text: String) {
+            label.text = text
+            label.textAlignment = .right
+            label.font = UIFont.systemFont(ofSize: 19, weight: .bold)
+        }
+
+    func arrangeStackView(
+        for stackView: UIStackView,
+        subviews: [UIView],
+        spacing: CGFloat = 0,
+        axis: NSLayoutConstraint.Axis = .horizontal,
+        distribution: UIStackView.Distribution = .fill,
+        aligment: UIStackView.Alignment = .fill
+    ) {
+        stackView.axis = axis
+        stackView.spacing = spacing
+        stackView.distribution = distribution
+        stackView.alignment = aligment
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        subviews.forEach { item in
+            item.translatesAutoresizingMaskIntoConstraints = false
+            stackView.addArrangedSubview(item)
+        }
+    }
+}
